+++ ml/pyproject.toml
@@ -9,6 +9,7 @@ authors = [
 requires-python = ">=3.13"
 dependencies = [
     "boto3>=1.35",
+    "flagembedding>=1.4.0",
     "httpx>=0.27",
     "igraph>=0.11",
     "leidenalg>=0.10",
@@ -16,6 +17,7 @@ dependencies = [
     "orjson>=3.10.7",
     "scikit-learn>=1.5",
     "scipy>=1.14",
+    "torch>=2.12.0",
 ]
 
 [build-system]
