+++ ml/src/pipeline/stages/evaluation/__init__.py
@@ -1,0 +1,3 @@
+from .main import run
+
+__all__ = ["run"]
